+++ components/common/NavBar/LargeNavBar/style.ts
