+++ components/common/NavBar/SmallNavBar/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function SmallNavBar() {
+  return <div></div>;
+}
+
+export default SmallNavBar;
