+++ components/common/NavBar/SmallNavBar/style.ts
@@ -1,0 +1,3 @@
+imsc;
+
+export {};
